chore: refactor adapters

+++ src/handlers/event-strategies/delete-event-strategy.ts
@@ -2,32 +2,33 @@ import { Event } from '../../@types/event'
 import { EventTags } from '../../constants/base'
 import { IEventRepository } from '../../@types/repositories'
 import { IEventStrategy } from '../../@types/message-handlers'
+import { IWebSocketAdapter } from '../../@types/adapters'
+import { WebSocketAdapterEvent } from '../../constants/adapter'
 
 
-export class DeleteEventStrategy implements IEventStrategy<Event, Promise<boolean>> {
+export class DeleteEventStrategy implements IEventStrategy<Event, Promise<void>> {
   public constructor(
+    private readonly webSocket: IWebSocketAdapter,
     private readonly eventRepository: IEventRepository,
   ) { }
 
-  public async execute(event: Event): Promise<boolean> {
-    try {
-      const eTags = event.tags.filter((tag) => tag[0] === EventTags.Event)
+  public async execute(event: Event): Promise<void> {
+    const eTags = event.tags.filter((tag) => tag[0] === EventTags.Event)
 
-      if (!eTags.length) {
-        return
-      }
-
-      await this.eventRepository.deleteByPubkeyAndIds(
-        event.pubkey,
-        eTags.map((tag) => tag[1])
-      )
-
+    if (!eTags.length) {
       return
-    } catch (error) {
-      console.error('Unable to handle event. Reason:', error)
-
-      return false
     }
+
+    const count = await this.eventRepository.create(event)
+    if (!count) {
+      return
+    }
+
+    await this.eventRepository.deleteByPubkeyAndIds(
+      event.pubkey,
+      eTags.map((tag) => tag[1])
+    )
+    this.webSocket.emit(WebSocketAdapterEvent.Broadcast, event)
   }
 
 }
